mirai-python-interface/requirements.txt
pyserial==3.5
pygame==2.5.2
matplotlib==3.8.0
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
scipy==1.11.3
opencv-python==4.8.1.78
pyside6==6.5.2
pyqtgraph==0.13.3